refactor(router): extract Router into RouteBuilder, RouteMatcher, and RouteDispatcher
Extract the 1632-line Router class into three focused classes:

- RouteBuilder: route construction, option validation, sub-route grouping
- RouteMatcher: URI resolution and matching against registered routes
- RouteDispatcher: dispatching matched routes to resources, middleware execution

Router retains its full public static API and delegates internally,
ensuring full backward compatibility.

Refs #375
